fix: Mejorar manejo de errores y precisión numérica
Correcciones críticas basadas en revisión de código:

**Crítico:**
- bot/handlers/user/favors.py: Cambiar int() a float() en _get_period_stats
  * Preserva valores fraccionarios de BesitoTransaction (ej: 0.1 por reacción)
  * Actualiza tipo de retorno a tuple[float, float]
  * Previene pérdida de datos en estadísticas

**Alta Prioridad - Error Logging:**
- bot/gamification/handlers/user/missions.py:
  * Agregar logger.warning en claim_mission_reward
  * Loguear errores al obtener racha y balance de usuario
  * Facilita debugging de problemas en servicios de gamificación

- bot/handlers/user/start.py:
  * Agregar logger.warning en _get_context_message
  * Loguear errores en verificación de misiones, favores y racha
  * Permite diagnosticar fallos en obtención de contexto dinámico

- bot/shop/handlers/user/shop.py:
  * Agregar logger.error en _get_user_favors
  * Loguear errores al obtener balance de favores
  * Previene mostrar saldo 0 incorrectamente sin diagnóstico

**Media Prioridad - Portabilidad DB:**
- alembic/versions/015_add_user_activity_tracking.py:
  * Usar sa.text('1') en lugar de '1' para server_default
  * Mejora portabilidad entre diferentes motores de base de datos
  * SQLAlchemy traduce correctamente el valor booleano

**Beneficios:**
✅ Cálculos precisos de Favores con valores fraccionarios
✅ Diagnóstico mejorado de errores en producción
✅ Mayor visibilidad de fallos en servicios
✅ Mejor portabilidad entre bases de datos
✅ Prevención de comportamiento silencioso incorrecto

Archivos modificados: 5
Líneas afectadas: ~15
Tests: Sintaxis validada ✓

+++ bot/handlers/user/favors.py
@@ -63,7 +63,7 @@ async def _get_period_stats(
     session: AsyncSession,
     user_id: int,
     start_date: datetime
-) -> tuple[int, int]:
+) -> tuple[float, float]:
     """Obtiene estadísticas de período (ganado/gastado).
 
     Args:
@@ -83,7 +83,7 @@ async def _get_period_stats(
         BesitoTransaction.created_at >= start_date
     )
     result = await session.execute(stmt_earned)
-    earned = int(result.scalar() or 0)
+    earned = float(result.scalar() or 0)
 
     # Gastado en el período (valor absoluto)
     stmt_spent = select(
@@ -94,7 +94,7 @@ async def _get_period_stats(
         BesitoTransaction.created_at >= start_date
     )
     result = await session.execute(stmt_spent)
-    spent = abs(int(result.scalar() or 0))
+    spent = abs(float(result.scalar() or 0))
 
     return earned, spent
